backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-google-genai
langchain-experimental
psycopg2-binary
python-dotenv
sqlalchemy
pydantic
